requirements.txt
playwright
playwright-stealth
resend
python-dotenv
